@@ -14,68 +14,61 @@
     "test:ci": "ng test --browsers ChromeHeadlessCI --code-coverage --watch=false",
     "test:chrome": "ng test --browsers Chrome --code-coverage --watch=true",
     "lint": "npm run lint:styles && ng lint --type-check",
-    "lint:styles": "stylelint \"src/app/**/*.scss\" \"src/assets/styles/**/*.scss\" --fix",
+    "lint:styles": "stylelint \"src/app/**/*.scss\" \"public/styles/**/*.scss\" --fix",
     "lint:ci": "ng lint",
     "lint:fix": "eslint -c .eslintrc.json --ext .ts src/ --fix",
     "commit": "git-cz"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.1.5",
-    "@angular/common": "^19.1.5",
-    "@angular/compiler": "^19.1.5",
-    "@angular/core": "^19.1.5",
-    "@angular/forms": "^19.1.5",
-    "@angular/material": "^19.1.3",
-    "@angular/platform-browser": "^19.1.5",
-    "@angular/platform-browser-dynamic": "^19.1.5",
-    "@angular/platform-server": "^19.1.5",
-    "@angular/router": "^19.1.5",
-    "@angular/ssr": "^19.1.6",
-    "@commitlint/cli": "^19.7.1",
-    "@ngx-translate/core": "^15.0.0",
-    "@ngx-translate/http-loader": "^8.0.0",
-    "cz-customizable": "^6.3.0",
+    "@angular/animations": "^19.2.14",
+    "@angular/cdk": "^19.2.16",
+    "@angular/common": "^19.2.0",
+    "@angular/compiler": "^19.2.0",
+    "@angular/core": "^19.2.0",
+    "@angular/forms": "^19.2.0",
+    "@angular/platform-browser": "^19.2.0",
+    "@angular/platform-browser-dynamic": "^19.2.0",
+    "@angular/platform-server": "^19.2.0",
+    "@angular/router": "^19.2.0",
+    "@angular/ssr": "^19.2.10",
+    "@commitlint/cli": "^19.8.0",
+    "cz-customizable": "^7.4.0",
+    "eslint": "^9.25.1",
+    "eslint-config-prettier": "^10.1.2",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-prettier": "^5.2.6",
+    "eslint-plugin-sonarjs": "^3.0.2",
+    "eslint-plugin-unicorn": "^59.0.0",
     "express": "^4.18.2",
+    "postcss": "^8.5.3",
+    "postcss-scss": "^4.0.9",
+    "prettier": "^3.5.3",
     "rxjs": "~7.8.0",
-    "stylelint": "^16.2.1",
-    "stylelint-config-property-sort-order-smacss": "^10.0.0",
-    "stylelint-declaration-block-no-ignored-properties": "^2.8.0",
-    "stylelint-no-unsupported-browser-features": "^8.0.1",
-    "stylelint-scss": "^6.1.0",
+    "stylelint": "^14.8.1",
+    "stylelint-config-property-sort-order-smacss": "^9.0.0",
+    "stylelint-config-standard-scss": "^3.0.0",
+    "stylelint-declaration-block-no-ignored-properties": "^2.5.0",
+    "stylelint-no-unsupported-browser-features": "^5.0.3",
+    "stylelint-prettier": "^2.0.0",
+    "stylelint-scss": "^4.2.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.1.6",
-    "@angular-eslint/builder": "^19.0.2",
-    "@angular-eslint/eslint-plugin": "^19.0.2",
-    "@angular-eslint/eslint-plugin-template": "^19.0.2",
-    "@angular-eslint/schematics": "^19.0.2",
-    "@angular-eslint/template-parser": "^19.0.2",
-    "@angular/cli": "^19.1.6",
-    "@angular/compiler-cli": "^19.1.5",
+    "@angular-devkit/build-angular": "^19.2.10",
+    "@angular/cli": "^19.2.10",
+    "@angular/compiler-cli": "^19.2.0",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
-    "@typescript-eslint/eslint-plugin": "^8.23.0",
-    "@typescript-eslint/parser": "^8.23.0",
-    "eslint": "^9.20.0",
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-prettier": "^5.2.3",
-    "eslint-plugin-sonarjs": "^3.0.1",
-    "eslint-plugin-unicorn": "^56.0.1",
-    "highlight.js": "^11.9.0",
-    "jasmine-core": "~5.1.0",
+    "jasmine-core": "~5.6.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "ngx-highlightjs": "^10.0.0",
-    "prettier": "^3.2.5",
-    "typescript": "~5.7.3"
+    "typescript": "~5.7.2"
   },
   "config": {
     "commitizen": {
