@@ -1,10 +1,14 @@
 import { mergeApplicationConfig, ApplicationConfig } from '@angular/core';
 import { provideServerRendering } from '@angular/platform-server';
-
+import { provideServerRouting } from '@angular/ssr';
 import { appConfig } from './app.config';
+import { serverRoutes } from './app.routes.server';
 
 const serverConfig: ApplicationConfig = {
-  providers: [provideServerRendering()]
+  providers: [
+    provideServerRendering(),
+    provideServerRouting(serverRoutes)
+  ]
 };
 
 export const config = mergeApplicationConfig(appConfig, serverConfig);
