@@ -1,14 +1,38 @@
-
 import { Component } from '@angular/core';
-import { MatTabsModule } from '@angular/material/tabs';
-
-import { IconsContributeComponent } from './icons-contribute/icons-contribute.component';
-import { IconsLibraryComponent } from './icons-library/icons-library.component';
-import { IconsUsageComponent } from './icons-usage/icons-usage.component';
+import { FormsModule } from '@angular/forms';
+import { IconComponent } from '@app/shared/components';
+import { IconName, IconSize } from '@app/shared/enums';
 
 @Component({
-    selector: 'app-icons',
-    imports: [MatTabsModule, IconsUsageComponent, IconsLibraryComponent, IconsContributeComponent],
-    templateUrl: './icons.component.html'
+  selector: 'app-icons',
+  imports: [IconComponent, FormsModule],
+  templateUrl: './icons.component.html',
+  styleUrl: './icons.component.scss'
 })
-export class IconsComponent {}
+export class IconsComponent {
+  public readonly IconName = IconName;
+  public readonly IconSize = IconSize;
+  public readonly iconNames = Object.values(IconName);
+
+  public searchTerm = '';
+  public selectedSize: IconSize = IconSize.M;
+  public copiedIcon: string | null = null;
+
+  protected get filteredIcons(): IconName[] {
+    return this.iconNames.filter((icon) => icon.toLowerCase().includes(this.searchTerm.toLowerCase()));
+  }
+
+  public async copyIconName(iconName: string): Promise<void> {
+    try {
+      await navigator.clipboard.writeText(iconName);
+      this.copiedIcon = iconName;
+
+      // Réinitialiser le message après 2 secondes
+      setTimeout(() => {
+        this.copiedIcon = null;
+      }, 2000);
+    } catch (err) {
+      console.error('Erreur lors de la copie :', err);
+    }
+  }
+}
