@@ -2,8 +2,10 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-spinner',
-  standalone: true,
+  imports: [],
   templateUrl: './spinner.component.html',
-  styleUrls: ['./spinner.component.scss']
+  styleUrl: './spinner.component.scss'
 })
-export class SpinnerComponent {}
+export class SpinnerComponent {
+
+}
