@@ -1,11 +1,7 @@
-import { ApplicationRef } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
-
 import { AppComponent } from './app/app.component';
 import { config } from './app/app.config.server';
 
-const bootstrap = (): Promise<ApplicationRef> => {
-  return bootstrapApplication(AppComponent, config);
-};
+const bootstrap = () => bootstrapApplication(AppComponent, config);
 
 export default bootstrap;
