@@ -1,7 +1,6 @@
 import { bootstrapApplication } from '@angular/platform-browser';
-
-import { AppComponent } from './app/app.component';
 import { appConfig } from './app/app.config';
+import { AppComponent } from './app/app.component';
 
-// eslint-disable-next-line no-undef
-bootstrapApplication(AppComponent, appConfig).catch((err) => console.error(err));
+bootstrapApplication(AppComponent, appConfig)
+  .catch((err) => console.error(err));
